requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
PyGithub==1.59.1
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.3
pytest==7.4.3
matplotlib==3.8.2
shapely==2.0.6
numpy>=1.24.0
pydantic>=2.7.0
jinja2>=3.1.0
jsonschema>=4.22.0
litellm>=1.77.0
openai>=1.0.0
mysql-connector-python>=8.0.0
